+++ CoinGecko/Entities/Response/Coins/SparklineIn7D.cs
@@ -5,6 +5,6 @@ namespace CoinGecko.Entities.Response.Coins
     public class SparklineIn7D
     {
         [JsonProperty("price")]
-        public double[] Price { get; set; }
+        public decimal[] Price { get; set; }
     }
 }
