+++ CoinGecko/Entities/Response/Simple/Price.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace CoinGecko.Entities.Response.Simple
 {
-    public class Price:Dictionary<string,Dictionary<string, double?>>
+    public class Price:Dictionary<string,Dictionary<string, decimal?>>
     {
     }
 }
